pyproject.toml
[project]
name = "django-order-management"
version = "0.1.0"
description = ""
authors = [
    {name = "shayanroychoudhury",email = "shayan.roy31@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-admin (>=2.0.2,<3.0.0)",
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg[binary] (>=3.2.11,<4.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
